styling modal
