+++ src/App/NavBar/index.js
@@ -1,29 +1,14 @@
 import React, { useState } from "react";
-import Modal from "../Modal";
+// import Modal from "../Modal";
 import "./index.css";
 
-function Nav() {
-  const [show, setModal] = useState(false);
-
-  function showModal() {
-    setModal(true);
-  }
-
-  function hideModal() {
-    setModal(false);
-  }
-
+function Nav({ children }) {
   return (
     <div className="nav">
       <div>
         <h1>Ferm.int</h1>
       </div>
-      <div>
-        <Modal show={show} handleClose={hideModal}>
-          <div>test</div>
-        </Modal>
-        <button onClick={showModal}>Start a ferment</button>
-      </div>
+      <div>{children}</div>
       <div>
         <label for="search">Search:</label>
         <input name="search"></input>
